pyproject.toml
[tool.poetry]
name = "pyfotobooth"
version = "0.1.0"
description = "Fotobooth project with backend frontend controller"
authors = ["DaMuBo <daniel910527@googlemail.com>"]
packages = [{include = "src", from = "."}]

[tool.poetry.scripts]
run_backend = "src.frontend.flask_app:app"

[tool.poetry.dependencies]
python = "^3.9"
segno = "^1.6.0"
pillow = "^10.1.0"
boto3 = "^1.34.9"
flask = "^3.0.0"
gunicorn = "^21.2.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
urllib3 = "1.26.18"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pylint = "^3.0.3"

[tool.poetry.group.prod]
optional = true

[tool.poetry.group.prod.dependencies]
pycups = "^2.0.1"
picamera = "^1.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

